Feat: Implementacion de la primera version del diseño de la vista home

+++ lib/models/models_api.dart
@@ -27,7 +27,7 @@ class Producto {
   factory Producto.fromMap(Map<String, dynamic> map) => Producto(
     id: '${map['id'] ?? ''}',
     title: (map['title']) ?? ' ',
-    price: (map['price']) ?? 0.0,
+    price: (map['price'] ?? 0.0).toDouble(),
     description: (map['description']) ?? ' ',
     category: (map['category']) ?? 'Unknown',
     image: (map['image']) ?? 'https://via.placeholder.com/150',
